backend/pyproject.toml
[project]
name = "bachata-buddy-api"
version = "1.0.0"
description = "REST API for Bachata Buddy choreography generation platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core Django
    "django>=5.2",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=5.3",
    # API Documentation
    "drf-spectacular>=0.27",
    "drf-spectacular[sidecar]>=0.27",
    # CORS Support
    "django-cors-headers>=4.3",
    # Filtering
    "django-filter>=23.0",
    # Database
    "psycopg2-binary>=2.9",
    # Image Processing
    "pillow>=10.0",
    # Google Cloud
    "google-cloud-run>=0.10",
    "google-cloud-storage>=2.10",
    "google-cloud-secret-manager>=2.16",
    # Production Server
    "gunicorn>=21.2",
    # Environment Variables
    "python-dotenv>=1.0",
    # Elasticsearch Client (for move search)
    "elasticsearch>=8.11",
    # Google AI (for natural language parsing)
    "google-generativeai>=0.3",
    # Video Processing Dependencies (compatible versions)
    "numpy>=1.24.0,<2.0.0",
    "librosa>=0.10.0",
    "opencv-python>=4.8.0,<4.10.0",
    "ultralytics>=8.0.0,<8.3.0",
    "scipy>=1.11.0,<1.14.0",
    # Audio Metadata
    "mutagen>=1.47.0",
    # Vector Search (FAISS for similarity search)
    "faiss-cpu>=1.7.4",
    "drf-spectacular-sidecar>=2025.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "api", "core"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
]
